+++ plugins/rp-epn/src/odc-rp-epn.cpp
@@ -20,31 +20,18 @@
 using namespace std;
 namespace bpo = boost::program_options;
 namespace bfs = boost::filesystem;
+namespace bpt = boost::property_tree;
 using namespace odc;
 using namespace odc::core;
 
-struct SResources
+struct SResource
 {
-    SResources(const string& _res)
+    SResource(const bpt::ptree& _pt)
     {
-        parse(_res);
-    }
-
-    string m_zone;
-    size_t m_n{ 0 };
-
-  private:
-    void parse(const string& _res)
-    {
-        boost::property_tree::ptree pt;
-        stringstream ss;
-        ss << _res;
-        boost::property_tree::read_json(ss, pt);
-
         // Do some basic validation of the input tree.
         // Only valid tags are allowed.
         set<string> validTags{ "zone", "n" };
-        for (const auto& v : pt)
+        for (const auto& v : _pt)
         {
             if (validTags.count(v.first.data()) == 0)
             {
@@ -53,9 +40,43 @@ struct SResources
                 throw runtime_error(ss.str());
             }
         }
-        m_zone = pt.get<string>("zone", "online");
-        m_n = pt.get<size_t>("n", 1);
+        m_zone = _pt.get<string>("zone", "online");
+        m_n = _pt.get<size_t>("n", 1);
     }
+
+    string m_zone;
+    size_t m_n{ 0 };
+};
+
+struct SResources
+{
+    SResources(const string& _res)
+    {
+        bpt::ptree pt;
+        stringstream ss;
+        ss << _res;
+        bpt::read_json(ss, pt);
+
+        try
+        {
+            // Single resource
+            m_resources.push_back(SResource(pt));
+        }
+        catch (const exception& _e)
+        {
+            // Array of resources
+            auto rpt{ pt.get_child_optional("") };
+            if (rpt)
+            {
+                for (const auto& v : rpt.get())
+                {
+                    m_resources.push_back(SResource(v.second));
+                }
+            }
+        }
+    }
+
+    vector<SResource> m_resources;
 };
 
 int main(int argc, char** argv)
@@ -67,6 +88,9 @@ int main(int argc, char** argv)
         bool release;
         CLogger::SConfig logConfig;
         partitionID_t partitionID;
+        string bash;
+        size_t numSlots{ 1 };
+        string wrkDir;
 
         // Generic options
         bpo::options_description options("odc-rp-epn options");
@@ -75,12 +99,18 @@ int main(int argc, char** argv)
         CCliHelper::addLogOptions(options, logConfig);
         CCliHelper::addOptions(options, partitionID);
         options.add_options()("res",
-                              bpo::value<string>(&res)->default_value("{\"zone\":\"online\",\"n\":10}"),
+                              bpo::value<string>(&res)->default_value("{\"zone\":\"online\",\"n\":1}"),
                               "Resource description in JSON format");
         options.add_options()(
             "host", bpo::value<string>(&host)->default_value("localhost:50001"), "EPN controller endpoint");
         options.add_options()(
             "release", bpo::bool_switch(&release)->default_value(false), "If set then release allocation");
+        options.add_options()("bash",
+                              bpo::value<string>(&bash)->default_value(""),
+                              "Optional bash commands to be prepended to SSH config file");
+        options.add_options()(
+            "slots", bpo::value<size_t>(&numSlots)->default_value(1), "Number of slots for each worker node");
+        options.add_options()("wrkdir", bpo::value<string>(&wrkDir)->default_value("/tmp/wn_dds"), "Working directory");
 
         // Parsing command-line
         bpo::variables_map vm;
@@ -108,26 +138,37 @@ int main(int argc, char** argv)
         if (!release)
         {
             auto r{ SResources(res) };
-            vector<string> nodes;
-            client.allocateNodes(partitionID, r.m_zone, r.m_n, nodes);
-            OLOG(ESeverity::info) << "epnc: nodes requested (" << r.m_n << ") allocated (" << nodes.size() << ")"
-                                  << endl;
+            dds::configRecords_t records;
+
+            for (const auto& v : r.m_resources)
+            {
+                vector<string> nodes;
+                client.allocateNodes(partitionID, v.m_zone, v.m_n, nodes);
+                OLOG(ESeverity::info) << "epnc: nodes requested (" << v.m_n << ") allocated (" << nodes.size() << ")"
+                                      << endl;
+
+                for (const auto& node : nodes)
+                {
+                    dds::configRecord_t record{ make_shared<dds::SConfigRecord>() };
+                    record->m_id = "wn_" + v.m_zone + "_" + node;
+                    record->m_addr = host;
+                    record->m_sshOptions = "";
+                    record->m_wrkDir = wrkDir;
+                    record->m_nSlots = numSlots;
+                    records.push_back(record);
+                }
+            }
 
             // Create SSH config file
-            // TODO: make hardcoded values as configuration parameters of the plugin
             const bfs::path tmpPath{ bfs::temp_directory_path() / bfs::unique_path() };
             bfs::create_directories(tmpPath);
             const bfs::path filepath{ tmpPath / "hosts.cfg" };
-            string sshOptions{ "" };
-            string wrkDir{ "/tmp/wn_dds" };
-            size_t numSlots{ 128 };
-            string bash{ "module load DataDistribution QualityControl" };
-            dds::CSSHConfigFile::make(filepath.string(), nodes, sshOptions, wrkDir, numSlots, bash);
+            dds::CSSHConfigFile::make(filepath.string(), records, bash);
 
             OLOG(ESeverity::info) << "epnc: SSH config file created at path " << filepath << endl;
 
             stringstream ss;
-            int requiredSlots(nodes.size() * numSlots);
+            int requiredSlots(records.size() * numSlots);
             ss << "<rms>ssh</rms><configFile>" << filepath.string() << "</configFile><requiredSlots>" << requiredSlots
                << "</requiredSlots>";