+++ core/src/ControlService.h
@@ -227,13 +227,17 @@ namespace odc::core
         {
         }
 
-        SActivateParams(const std::string& _topologyFile, const std::string& _topologyContent)
+        SActivateParams(const std::string& _topologyFile,
+                        const std::string& _topologyContent,
+                        const std::string& _topologyScript)
             : m_topologyFile(_topologyFile)
             , m_topologyContent(_topologyContent)
+            , m_topologyScript(_topologyScript)
         {
         }
         std::string m_topologyFile;    ///< Path to the topoloy file
         std::string m_topologyContent; ///< Content of the XML topology
+        std::string m_topologyScript;  ///< Script that generates topology content
 
         // \brief ostream operator.
         friend std::ostream& operator<<(std::ostream& _os, const SActivateParams& _params);
@@ -246,13 +250,17 @@ namespace odc::core
         {
         }
 
-        SUpdateParams(const std::string& _topologyFile, const std::string& _topologyContent)
+        SUpdateParams(const std::string& _topologyFile,
+                      const std::string& _topologyContent,
+                      const std::string& _topologyScript)
             : m_topologyFile(_topologyFile)
             , m_topologyContent(_topologyContent)
+            , m_topologyScript(_topologyScript)
         {
         }
         std::string m_topologyFile;    ///< Path to the topoloy file
         std::string m_topologyContent; ///< Content of the XML topology
+        std::string m_topologyScript;  ///< Script that generates topology content
 
         // \brief ostream operator.
         friend std::ostream& operator<<(std::ostream& _os, const SUpdateParams& _params);