+++ ReleaseNotes.md
@@ -1,6 +1,8 @@
 # ODC Release Notes
 
 ## v0.38 (NOT YET RELEASED)
+### ODC common
+Added: new `content` parameter to `Activate`, `Update` and `Run` requests. One can set either a topology filepath or content.    
 
 
 ## v0.36 (2021-08-13)