+++ ReleaseNotes.md
@@ -1,6 +1,9 @@
 # ODC Release Notes
 
-## v0.38 (NOT YET RELEASED)
+## v0.40 (NOT YET RELEASED)
+
+
+## v0.38 (2021-09-01)
 ### ODC common
 Added: new `content` and `script` parameters to `Activate`, `Update` and `Run` requests. One can set either a topology filepath, content or shell commands. If `content` is set than ODC creates a temp topology file with that content. If `script` is set than ODC executes the script and saves stdout to a temp topology file.    
 Added: `odc-rp-epn` plugin supports array of resources.